chore: move directory of inventory/change-history
